#58 | Moved the clock model
